version: use alternative baby proof measure for version file.
the `.gitignore` file is used by build system in non intuitive ways that
makes it difficult to ship the buildinfo file.

Even when shipped as part of the distribute, pip will ignore the
file when installing the package due to gitignore.

This change removes the buildinfo file from the gitignore, and
instead uses a test to catch erronous buildinfo commits to our
VCS.